test(cli): define what is failing
