+++ packages/cli/test/setup.js
@@ -1,3 +1,4 @@
 // ease output comparison
 // listr renders differently if isTTY is not true
 process.stdout.isTTY = undefined
+jest.setTimeout(10000)
